backend/requirements.txt
crewai
litellm
python-dotenv
uvicorn
fastapi
google-generativeai
